refactored packages to make more sense with the fragment activity
