+++ demo-app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.shop.AstronomyShopActivity"
+            android:name=".shop.ui.AstronomyShopActivity"
             android:exported="true"
             android:label="@string/title_activity_astronomy_shop"
             />
