+++ demo-app/src/main/java/io/opentelemetry/android/demo/shop/ui/products/ProductList.kt
@@ -1,4 +1,4 @@
-package io.opentelemetry.android.demo.ui.shop.products
+package io.opentelemetry.android.demo.shop.ui.products
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
@@ -8,7 +8,7 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import io.opentelemetry.android.demo.model.Product
+import io.opentelemetry.android.demo.shop.model.Product
 
 @Composable
 fun ProductList(products: List<Product>, onProductClick: (String) -> Unit) {
